# [Minor] Speedup to_array_of_size for Decimal128